No support for image with namespace and tag
I encountered an issue where if I was attempting
run a container from an image that was in a private
repository, and the image was prefixed with a namespace
and a tag is provided, then it would fail.

I have added a test for this and updated the image name
parser to accept an arbitrary number of "sub-directories"
as namespaces.  This doesn't appear to be true to the
spec but perhaps some larger organizations may divide
their images by organization/project, e.g.

example.registry.com:5000/org/team/image:1.0